[Feature] - Creating the UserController to create new users
