+++ src/main/java/com/example/api/dealership/adapter/entrypoint/AuthenticationController.java
@@ -2,17 +2,16 @@ package com.example.api.dealership.adapter.entrypoint;
 
 import com.example.api.dealership.adapter.dtos.Response;
 import com.example.api.dealership.adapter.dtos.user.UserDtoRequest;
-
-import com.example.api.dealership.adapter.mapper.UserMapper;
 import com.example.api.dealership.adapter.service.security.AuthenticationService;
-import com.example.api.dealership.adapter.service.user.UserService;
 import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 @Slf4j
 @RestController
@@ -20,13 +19,7 @@ import org.springframework.web.bind.annotation.*;
 @RequiredArgsConstructor
 public class AuthenticationController {
 
-    private final UserService userService;
-
-    private final PasswordEncoder passwordEncoder;
-
     private final AuthenticationService authenticationService;
-
-    private final UserMapper userMapper;
 
     @Operation(summary = "Authenticate a valid user")
     @PostMapping(path = "/auths",produces = "application/json")
@@ -44,29 +37,6 @@ public class AuthenticationController {
         }
     }
 
-    @Operation(summary = "Create a user")
-    @PostMapping(path = "/users",produces = "application/json")
-    public ResponseEntity<Response<String>> saveUser(@RequestBody UserDtoRequest userDtoRequest){
 
-        var response = new Response<String>();
-
-        var optinalUser = userService.findByUsername(userDtoRequest.getUsername());
-
-        if (optinalUser.isPresent()) {
-            response.setData("This username already exists.");
-
-            return ResponseEntity.badRequest().body(response);
-        }
-
-        var userModel = userMapper.toUserModel(userDtoRequest);
-
-        userModel.setPassword(passwordEncoder.encode(userModel.getPassword()));
-
-        var user = userService.saveUser(userModel);
-
-        response.setData("User: " + user.getUsername() + " created with success");
-
-        return ResponseEntity.ok().body(response);
-    }
 
 }
